Update uploadcert.php
add check exist
add write record to db

+++ forms/uploadcert.php
@@ -19,21 +19,35 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
                 $hash = md5_file($file["tmp_name"]);
 
                 $dir = "/web/uploads/"; // modify this accourding to your server's upload directory
-                
-                $path = $dir . date("Y-m-d") . "/";
-                if(is_dir($path)){
+                $date = date("Y-m-d");
+                $path = $dir . $date . "/";
+                if (is_dir($path)) {
                     //echo "path exists";
-                }else{
+                } else {
                     //echo "path not exists";
                     mkdir($path, 0777, true);
                 }
                 $newfile = $path . $hash . "." . $filetype[1];
 
-                if (move_uploaded_file($file["tmp_name"], $newfile)){
-                    echo "successfully uploaded your certificate";
-                }else{
-                    echo "failed to uploaded your certificate, please try again";
+                // check if the file is already exists in both directory and db
+                $fileinfo = dbsearchmultiplecondition("image", ["imageid"], array("date" => $date, "image" => $hash));
+                if (file_exists($newfile) or count($fileinfo) > 0) {
+                    echo "File already exists";
+                } else {
+                    if (move_uploaded_file($file["tmp_name"], $newfile)) {
+                        $values['worker'] = $_SESSION['uid'];
+                        $values['date'] = "'" . $date . "'";
+                        $values['image'] = "'" . $hash . "'";
+                        $values['type'] = "'" . $filetype[1] . "'";
+                        if (dbinsert("image", $values)) {
+                            echo "successfully uploaded your certificate";
+                        }
+                    } else {
+                        echo "failed to uploaded your certificate, please try again";
+                    }
                 }
+
+
             }
         }
     }
